Add tests and tidy up files
