+++ pybam/__init__.py
@@ -1,1 +1,3 @@
-from .pybam import *
+from .defbamentry import *
+from .defbamfile import *
+from .defexceptions import *
