+++ pybam/defexceptions.py
@@ -1,0 +1,3 @@
+# custom exceptions
+class PybamWarn(Exception): pass
+class PybamError(Exception): pass
